We can use references in structs now!!
