+++ test/test.rs
@@ -2,10 +2,8 @@
 
 extern crate crdts;
 
-mod gset;
 mod lwwreg;
 mod map;
 mod mvreg;
 mod orswot;
-mod pncounter;
 mod vclock;
